requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==37.0.1
cycler==0.11.0
Deprecated==1.2.13
depthai==2.15.3.0
-e git+https://github.com/luxonis/depthai.git@0a96dbd8bb44b1aa991dcda2025248a5e7c2aff0#egg=depthai_sdk&subdirectory=depthai_sdk
fastapi==0.75.1
ffmpy3==0.2.4
fonttools==4.32.0
gcloud==0.18.3
google-auth==2.6.5
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.0
grpcio==1.44.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
jwcrypto==1.2
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
opencv-contrib-python==4.5.4.58
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pyparsing==3.0.8
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jwt==3.3.2
python-multipart==0.0.5
pytube==12.0.0
PyTurboJPEG==1.6.4
pytz==2022.1
pyusb==1.2.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.2
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.14.1
